refactor(phase0): move the clippy policy to the workspace
Modularization plan, Phase 0. `clippy::all` + `clippy::pedantic` previously
existed only as crate attributes on `backend/src/main.rs`, so any module moved
out of the backend into a new crate would silently lose them. The policy now
lives in `[workspace.lints.clippy]` and every member crate opts in with
`[lints] workspace = true`, which is a prerequisite for extracting anything.

No behaviour or public configuration changes.

The backend's five justified allows (module_name_repetitions,
must_use_candidate, return_self_not_must_use, missing_errors_doc,
large_futures) move to the workspace table verbatim, so the set of warnings the
backend emits is unchanged.

Extending the policy to `shared`, `frontend`, the `flags_builder` bin and the
integration tests surfaced 906 warning sites that had never been linted.
`cargo clippy --fix` was run to convergence and cleared 498 of them; the
remaining 408 have no machine-applicable fix. They are handled as follows:

- backend (7 sites) is fixed or narrowly justified per site and keeps the full
  policy with no crate-level allow list. `flags_builder` is a build-time asset
  packer whose narrowing casts are bounded by the checked-in SVG set;
  `provider_dns_https_sni`'s manual Debug impl now uses
  `finish_non_exhaustive()`.
- `shared` and `frontend` carry documented crate-level allow lists in their
  lib.rs. Clearing them needs hand edits to DTO/serde code and to Yew `html!`
  bodies, which is behaviour-bearing work that Phase 0 must not mix in. The
  lists are tracked debt scoped to those crates: crates extracted from the
  backend inherit the strict policy, not this one.
- four sites in `shared` keep an explicit `#[allow]` because clippy's own
  suggestion names a private module and does not compile.

Gates: nightly rustfmt, nightly clippy `-D warnings`, `cargo test --workspace`
and markdownlint are green.

+++ shared/src/model/pagination.rs
@@ -33,7 +33,7 @@ impl PageRequestDto {
         }
     }
 
-    /// Clamps page_size to the allowed range [25, 200] and snaps to nearest valid value.
+    /// Clamps `page_size` to the allowed range [25, 200] and snaps to nearest valid value.
     pub fn normalize_page_size(&mut self) {
         let mut size = self.page_size;
         if size < VALID_PAGE_SIZES[0] {
@@ -71,7 +71,8 @@ impl<T: PartialEq> PagedResponseDto<T> {
     /// Constructs a paged response from a full items vec and pagination parameters.
     /// `all_items` is the complete filtered+aggregated result set (already sorted).
     pub fn new(items: Vec<T>, page: u32, page_size: u16, total_items: u64) -> Self {
-        let total_pages = if total_items == 0 { 0 } else { ((total_items as f64) / (page_size as f64)).ceil() as u32 };
+        let total_pages =
+            if total_items == 0 { 0 } else { ((total_items as f64) / f64::from(page_size)).ceil() as u32 };
         let has_prev = page > 1;
         let has_next = total_items > 0 && page < total_pages;
         Self { items, page, page_size, total_items, total_pages, has_prev, has_next }
@@ -82,13 +83,13 @@ impl<T: PartialEq> PagedResponseDto<T> {
         if self.total_items == 0 {
             0
         } else {
-            ((self.page - 1) * self.page_size as u32) as u64 + 1
+            u64::from((self.page - 1) * u32::from(self.page_size)) + 1
         }
     }
 
     /// Returns the 1-based end index of the current page.
     pub fn range_end(&self) -> u64 {
-        let end = (self.page as u64) * self.page_size as u64;
+        let end = u64::from(self.page) * u64::from(self.page_size);
         end.min(self.total_items)
     }
 }
